chore: setup as eks microservice
