+++ .env
@@ -1,0 +1,4 @@
+HOST_PORT_SERVER=8080
+CONTAINER_PORT_SERVER=8080
+PORT_SERVER=8080
+VOLUME_NODE_API=/home/node/app
