+++ .github/workflows/microservice-build.yaml
@@ -1,9 +1,9 @@
-name: Microservice - Continuous Integration / Build
+name: CI - Microservice Build & Test
 
 on:
   push:
     branches:
-      - "*"
+      - "main"
   pull_request:
     types:
       - opened
@@ -14,15 +14,19 @@ jobs:
   build:
     runs-on: ubuntu-latest
 
-    # strategy:
-    #   matrix:
-        # node-version: [20.x]
+    strategy:
+      matrix:
+        node-version: [20.x]
 
     steps:
       - uses: actions/checkout@v4
 
-      # INSERT STEPS RELATED TO PROJECT BUILD AND TEST
-      
+      - name: Install Dependencies
+        run: npm ci
+
+      - name: Build Project
+        run: npm run build --if-present
+
       - name: SonarQube Scan
         uses: SonarSource/sonarqube-scan-action@v4
         env:
