+++ infra/aws/provider.tf
@@ -1,0 +1,3 @@
+provider "aws" {
+  region = var.regionDefault
+}
