+++ infra/aws/vars.tf
@@ -1,0 +1,7 @@
+variable "regionDefault" {
+  default = "us-east-1"
+}
+
+variable "repositoryName" {
+  default = "tech-challenge-lanchonete/api-docs"
+}
