+++ infra/chart-values/aws-values.yaml
@@ -1,0 +1,20 @@
+configMapData:
+  SERVERS: '[{ "url": "http://a08a6972914de4700a523822e077d3ad-1975528089.us-east-1.elb.amazonaws.com", "description": "AWS EKS Cluster"}]'
+
+image:
+  repository: "045628470274.dkr.ecr.us-east-1.amazonaws.com/tech-challenge-lanchonete/api-docs"
+  tag: "latest"
+
+service:
+  type: ClusterIP
+  port: 8080
+  targetPort: 8080
+
+replicaCount: 1
+resources:
+  requests:
+    memory: "64Mi"
+    cpu: "100m"
+  limits:
+    memory: "128Mi"
+    cpu: "200m"
